+++ ema.cabal
@@ -156,6 +156,7 @@ library
     Ema.Route.Generic.SubModel
     Ema.Route.Generic.TH
     Ema.Route.Generic.Verification
+    Ema.Route.Generic.Verification.Any
     Ema.Route.Lib.File
     Ema.Route.Lib.Folder
     Ema.Route.Lib.Multi
